+++ packages/engine/src/helpers/useDecisionCallback.ts
@@ -45,14 +45,12 @@ export default function useDecisionCallback({
     // FIXME: Verificar a session e o payload
     // Está enviando session vazia quando no primeiro passo.
     // quando em outros passos, não pega o payload atual que foi enviado.
-    console.log("hehe")
     if (decisionCallbackStrategy === "external") {
       if (webhookConfig && stepWebhookDefinition) {
         callWebhook(webhookConfig, {
           sessionData: session,
           receivingStepData: payload,
         }).then((webhookResponse) => {
-          console.log("lulul,", webhookResponse)
           res(webhookResponse)
         })
       }
