+++ packages/engine/src/mocks/sample-experience.js
@@ -167,28 +167,6 @@ const mock = {
   ],
 }
 
-/** @type {IExperienceJSONSchema} */
-const sampleWithWebhookConfig = {
-  "webhookConfig": {
-    "triggers": [
-      {
-        "slug": "personal_data",
-        "saveProcessedData": true,
-      },
-      {
-        "slug": "documents",
-        "saveProcessedData": true,
-      },
-    ],
-    "url": "https://teste.com.br",
-    "headers": {
-      "Content-Type": "application/json",
-      "Authorization": "Bearer ",
-    },
-  },
-  ...mock,
-}
-
 module.exports = mock
 
-exports.sampleWithWebhookConfig = sampleWithWebhookConfig
+
